Change consumer and provider to classes
